requirements/base.txt
asgiref==3.6.0
boto3==1.20.46
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.10.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Pillow==8.4.0
pytz==2022.2.1
six==1.16.0
psycopg2-binary==2.9.2   
django-debug-toolbar==3.8.1
PyYAML==6.0.1
python-telegram-bot
qrcode[pil]
python-dotenv
djongo
pymongo==3.11.0

